requirements.txt
matplotlib==3.7.2
Pillow==10.0.1
Pillow==10.1.0
pynput==1.7.6
